+++ hopsworksai/provider_test.go
@@ -1,10 +1,13 @@
 package hopsworksai
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"strings"
 	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -199,4 +202,113 @@ func testAccResourceDataSourceCheckAllAttributes(resourceName string, dataSource
 		}
 		return nil
 	}
+}
+
+func TestProviderAPIGateway(t *testing.T) {
+	testCases := []struct {
+		config              *terraform.ResourceConfig
+		expectedGateway     string
+		expectedDiagnostics diag.Diagnostics
+	}{
+		{
+			config: terraform.NewResourceConfigRaw(map[string]interface{}{
+				"api_gateway": "https://api.gateway.ai",
+			}),
+			expectedGateway: "https://api.gateway.ai",
+			expectedDiagnostics: diag.Diagnostics{
+				ApiGatewayDevDiagnostic,
+			},
+		},
+
+		{
+			config:          terraform.NewResourceConfigRaw(map[string]interface{}{}),
+			expectedGateway: api.DEFAULT_API_GATEWAY,
+		},
+
+		{
+			config: terraform.NewResourceConfigRaw(map[string]interface{}{
+				"api_gateway": "api.gateway.ai",
+			}),
+			expectedDiagnostics: diag.Diagnostics{
+				ApiGatewayDevDiagnostic,
+				{
+					Severity: diag.Error,
+					Summary:  "API Gateway URL is missing scheme http/https",
+				},
+			},
+		},
+
+		{
+			config: terraform.NewResourceConfigRaw(map[string]interface{}{
+				"api_gateway": ":foo",
+			}),
+			expectedDiagnostics: diag.Diagnostics{
+				ApiGatewayDevDiagnostic,
+				{
+					Severity: diag.Error,
+					Summary:  "Could not parse API Gateway URL",
+				},
+			},
+		},
+	}
+
+	var provider *schema.Provider
+	for _, test := range testCases {
+		provider = Provider("dev")()
+		diagnostics := provider.Validate(test.config)
+		apiGatewayInTest, found := test.config.Get("api_gateway")
+		if !found {
+			apiGatewayInTest = api.DEFAULT_API_GATEWAY
+		}
+		t.Log(fmt.Sprintf("Testing case api_gateway: %s", apiGatewayInTest))
+		if len(test.expectedDiagnostics) > 0 {
+			if test.expectedDiagnostics.HasError() && !diagnostics.HasError() {
+				t.Fatalf("Expected Validate to return Diagnostic error %s but it did not",
+					diagnosticsSummary(test.expectedDiagnostics))
+			}
+			if !validateDiagnosticErrors(test.expectedDiagnostics, diagnostics) {
+				t.Fatalf("Expected to find Diagnostics summary \"%s\" in Validation but found \"%s\"",
+					diagnosticsSummary(test.expectedDiagnostics), diagnosticsSummary(diagnostics))
+			}
+		} else {
+			provider.Configure(context.Background(), test.config)
+			c, ok := provider.Meta().(*api.HopsworksAIClient)
+			if !ok {
+				t.Error("client is not HopsworksAIClient")
+			}
+			if c.ApiGateway != test.expectedGateway {
+				t.Errorf("Expected API Gateway to be %s but it is %s", test.expectedGateway, c.ApiGateway)
+			}
+		}
+	}
+}
+
+func diagnosticsSummary(diagnostics diag.Diagnostics) string {
+	var summary strings.Builder
+	for i, d := range diagnostics {
+		summary.WriteString(d.Summary)
+		if i < len(diagnostics)-1 {
+			summary.WriteString(" - ")
+		}
+	}
+	return summary.String()
+}
+
+func validateDiagnosticErrors(expected, actual diag.Diagnostics) bool {
+	if len(expected) != len(actual) {
+		return false
+	}
+	for _, expectedDiag := range expected {
+		var diagnosticFound bool
+		for _, actualDiag := range actual {
+			if expectedDiag.Summary == actualDiag.Summary {
+				diagnosticFound = true
+				break
+			}
+		}
+		if !diagnosticFound {
+			return false
+		}
+	}
+	return true
 }